requirements.txt
bert-score>=0.3.13
rouge-score>=0.1.2
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.0
pandas>=2.1.3
numpy>=1.24.3
scikit-learn>=1.3.2
streamlit>=1.32.0
youtube-transcript-api>=0.6.2 